+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoModule.kt
@@ -1,10 +1,8 @@
 package expo.modules.video
 
 import android.app.Activity
-import android.view.View
 import androidx.media3.common.MediaItem
 import androidx.media3.common.PlaybackParameters
-import androidx.media3.ui.DefaultTimeBar
 import expo.modules.kotlin.exception.Exceptions
 import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.modules.ModuleDefinition
@@ -34,24 +32,22 @@ class VideoModule : Module() {
         view.playerView.resizeMode = contentFit.toResizeMode()
       }
 
+      AsyncFunction("enterFullscreen") { view: VideoView ->
+        view.enterFullscreen()
+      }
+
+      Prop("allowsFullscreen") { view: VideoView, allowsFullscreen: Boolean? ->
+        view.allowsFullscreen = allowsFullscreen ?: true
+      }
+
       Prop("requiresLinearPlayback") { view: VideoView, requiresLinearPlayback: Boolean? ->
         val linearPlayback = requiresLinearPlayback ?: false
-        view.playerView.setShowFastForwardButton(!linearPlayback)
-        view.playerView.setShowRewindButton(!linearPlayback)
-        view.playerView.setShowPreviousButton(!linearPlayback)
-        view.playerView.setShowNextButton(!linearPlayback)
-
-        // TODO: Make the requiresLinearPlayback hide only the scrubber instead of the whole progress bar.
-        //  Maybe use custom layout for the player as the scrubber is not available?
-        val progressBar = view.playerView.findViewById<View>(androidx.media3.ui.R.id.exo_progress)
-        if (progressBar is DefaultTimeBar) {
-          progressBar.visibility = if (linearPlayback) {
-            View.GONE
-          } else {
-            View.VISIBLE
-          }
-        }
+        view.playerView.applyRequiresLinearPlayback(linearPlayback)
         view.videoPlayer?.requiresLinearPlayback = linearPlayback
+      }
+
+      OnViewDestroys {
+        VideoViewManager.removeVideoView(it.id)
       }
     }